`SwipeActionsView.Style` -> `SwipeActionsViewStyle`
